extension: add support for locally published controller image
Previously, it was easy to have the sample extension install use a local
image, because the `source` field could be set. But having the
extension's controller deployment use the local image was a bit
trickier, because the deployment needs to work well for local
development and external publishing. This changeset addresses the local
development use-case by adding some `make` recipes for patching the
extension's `install.yaml` with a local image for local development.

Signed-off-by: suskin <813571+suskin@users.noreply.github.com>