controller: add wordpress instance uniqueness
Since a stack may need to be able to create multiple instances (of
wordpress in this case), those instances should be unique enough that N
of them can be created. This changeset adds that functionality by using
the UID of the wordpress instance CR created by the user, and including
that in all of the created resources' names. It also adds some light
tagging for tracking purposes, so that objects can be traced back to
having been created by this stack.

This changeset also adds a pass at a complete but shallow guide for how
to try out the stack.

Signed-off-by: suskin <813571+suskin@users.noreply.github.com>